# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AMIS API
  description: Advanced Medical Information System API
  version: 1.0.0
  contact:
    name: AMIS Support
    email: support@amis.uz
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.amis.uz/api/v1
    description: Production server
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /patients:
    get:
      tags: [Patients]
      summary: List patients
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: search
          in: query
          schema:
            type: string
          description: Search by name, phone, or ID
        - name: gender
          in: query
          schema:
            type: string
            enum: [male, female]
        - name: citizenship
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Patient'
    post:
      tags: [Patients]
      summary: Create patient
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientInput'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
  /patients/{id}:
    get:
      tags: [Patients]
      summary: Get patient by ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Patient details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Patients]
      summary: Update patient
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientInput'
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
...
  /appointments:
    get:
      tags: [Appointments]
      summary: List appointments
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: patient_id
          in: query
          schema:
            type: string
            format: uuid
        - name: doctor_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, waiting, in_progress, completed, cancelled]
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of appointments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Appointment'
    post:
      tags: [Appointments]
      summary: Create appointment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentInput'
      responses:
        '201':
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
...
  /analytics/dashboard:
    get:
      tags: [Analytics]
      summary: Get dashboard data
      security:
        - BearerAuth: []
      parameters:
        - name: clinic_id
          in: query
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
...
  /references/services:
    get:
      tags: [References]
      summary: Get services catalog
      security:
        - BearerAuth: []
      parameters:
        - name: group_id
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Services catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    PaginatedResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'
...
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clinic_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        patronymic:
          type: string
        birth_date:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female]
        phone:
          type: string
        phone_2:
          type: string
        email:
          type: string
        citizenship:
          type: string
        address:
          type: string
        deposit_balance:
          type: number
          format: float
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    PatientInput:
      type: object
      required: [first_name, last_name, birth_date, gender, phone]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        patronymic:
          type: string
        birth_date:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female]
        phone:
          type: string
        phone_2:
          type: string
        email:
          type: string
        citizenship:
          type: string
        address:
          type: string
        passport:
          type: string
        price_category_id:
          type: string
          format: uuid
        notes:
          type: string
    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clinic_id:
          type: string
          format: uuid
        branch_id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        doctor_id:
          type: string
          format: uuid
        service_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [scheduled, waiting, in_progress, completed, cancelled]
        appointment_date:
          type: string
          format: date
        start_time:
          type: string
        end_time:
          type: string
        booking_method:
          type: string
        cabinet:
          type: string
        notes:
          type: string
        patient:
          $ref: '#/components/schemas/Patient'
        doctor:
          $ref: '#/components/schemas/Staff'
        service:
          $ref: '#/components/schemas/Service'
    AppointmentInput:
      type: object
      required: [patient_id, doctor_id, service_id, appointment_date, start_time]
      properties:
        patient_id:
          type: string
          format: uuid
        doctor_id:
          type: string
          format: uuid
        service_id:
          type: string
          format: uuid
        appointment_date:
          type: string
          format: date
        start_time:
          type: string
        booking_method:
          type: string
          enum: [manual, online, referral]
        referral_doctor_id:
          type: string
          format: uuid
        contract_id:
          type: string
          format: uuid
        cabinet:
          type: string
        notes:
          type: string
...
    Staff:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        patronymic:
          type: string
        specialty:
          type: string
        position:
          type: string
        cabinet:
          type: string
        photo_url:
          type: string
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        duration:
          type: integer
        base_price:
          type: number
          format: float
        group_id:
          type: string
          format: uuid
...
    Dashboard:
      type: object
      properties:
        total_appointments:
          type: integer
        completed_appointments:
          type: integer
        cancelled_appointments:
          type: integer
        total_revenue:
          type: number
          format: float
        new_patients:
          type: integer
        waiting_patients:
          type: integer
        by_service_group:
          type: object
        by_doctor:
          type: array
          items:
            type: object
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
        clinic_id:
          type: string
          format: uuid
...